+++ projects/core/src/checkout/checkout.module.ts
@@ -3,7 +3,7 @@ import { NgModule } from '@angular/core';
 import { CheckoutService } from './facade/index';
 import { CheckoutStoreModule } from './store/checkout-store.module';
 import { PageTitleResolver } from '../cms/index';
-import { CheckoutPageTitleResolver } from './service/checkout-page-title.resolver';
+import { CheckoutPageTitleResolver } from './services/checkout-page-title.resolver';
 
 @NgModule({
   imports: [CheckoutStoreModule],
